+++ pkg/types/entitlement/entilement.go
@@ -1,6 +1,8 @@
 package entitlement
 
 import (
+	"fmt"
+
 	v2 "github.com/conductorone/baton-sdk/pb/c1/connector/v2"
 	"github.com/conductorone/baton-sdk/pkg/annotations"
 	"google.golang.org/protobuf/proto"
@@ -30,8 +32,42 @@ func WithDisplayName(displayName string) EntitlementOption {
 	}
 }
 
-func WithDescriptionName(description string) EntitlementOption {
+func WithDescription(description string) EntitlementOption {
 	return func(g *v2.Entitlement) {
 		g.Description = description
 	}
+}
+
+func NewEntitlementID(resource *v2.Resource, permission string) string {
+	return fmt.Sprintf("%s:%s:%s", resource.Id.ResourceType, resource.Id.Resource, permission)
+}
+
+func NewPermissionEntitlement(resource *v2.Resource, name string, entitlementOptions ...EntitlementOption) *v2.Entitlement {
+	entitlement := &v2.Entitlement{
+		Id:          NewEntitlementID(resource, name),
+		DisplayName: name,
+		Slug:        name,
+		Purpose:     v2.Entitlement_PURPOSE_VALUE_PERMISSION,
+		Resource:    resource,
+	}
+
+	for _, entitlementOption := range entitlementOptions {
+		entitlementOption(entitlement)
+	}
+	return entitlement
+}
+
+func NewAssignmentEntitlement(resource *v2.Resource, name string, entitlementOptions ...EntitlementOption) *v2.Entitlement {
+	entitlement := &v2.Entitlement{
+		Id:          NewEntitlementID(resource, name),
+		DisplayName: name,
+		Slug:        name,
+		Purpose:     v2.Entitlement_PURPOSE_VALUE_ASSIGNMENT,
+		Resource:    resource,
+	}
+
+	for _, entitlementOption := range entitlementOptions {
+		entitlementOption(entitlement)
+	}
+	return entitlement
 }
