+++ pkg/types/resource/resource_test.go
@@ -1,12 +1,11 @@
-package sdk
+package resource
 
 import (
 	"testing"
 
 	v2 "github.com/conductorone/baton-sdk/pb/c1/connector/v2"
 	"github.com/conductorone/baton-sdk/pkg/annotations"
-	eopt "github.com/conductorone/baton-sdk/pkg/types/entitlement"
-	ropt "github.com/conductorone/baton-sdk/pkg/types/resource"
+	sdk "github.com/conductorone/baton-sdk/pkg/sdk"
 	"github.com/stretchr/testify/require"
 )
 
@@ -15,7 +14,7 @@ func TestNewAppResource(t *testing.T) {
 		"app_name": "Test",
 	}
 	rt := NewResourceType("App", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_APP})
-	ar, err := NewAppResource("test app", rt, nil, 1234, "https://example.com", profile, ropt.WithAnnotation(&v2.V1Identifier{Id: "v1"}))
+	ar, err := NewAppResource("test app", rt, nil, 1234, "https://example.com", profile, WithAnnotation(&v2.V1Identifier{Id: "v1"}))
 	require.NoError(t, err)
 	require.NotNil(t, ar)
 	require.Equal(t, rt.Id, ar.Id.ResourceType)
@@ -29,91 +28,20 @@ func TestNewAppResource(t *testing.T) {
 	require.True(t, ok)
 	require.Equal(t, "v1", v1ID.Id)
 
-	roleTrait, err := GetRoleTrait(ar)
+	roleTrait, err := sdk.GetRoleTrait(ar)
 	require.Error(t, err)
 	require.Nil(t, roleTrait)
-	appTrait, err := GetAppTrait(ar)
+	appTrait, err := sdk.GetAppTrait(ar)
 	require.NoError(t, err)
 	require.NotNil(t, appTrait)
 	require.Equal(t, "https://example.com", appTrait.HelpUrl)
 	require.NotNil(t, appTrait.Profile)
-	fName, foundProfileData := GetProfileStringValue(appTrait.Profile, "app_name")
+	fName, foundProfileData := sdk.GetProfileStringValue(appTrait.Profile, "app_name")
 	require.True(t, foundProfileData)
 	require.Equal(t, "Test", fName)
-	mName, foundProfileData := GetProfileStringValue(appTrait.Profile, "first_name")
+	mName, foundProfileData := sdk.GetProfileStringValue(appTrait.Profile, "first_name")
 	require.False(t, foundProfileData)
 	require.Equal(t, "", mName)
-}
-
-func TestNewAssignmentEntitlement(t *testing.T) {
-	rt := NewResourceType("Group", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_GROUP})
-	ur, err := NewResource("test-group", rt, nil, 1234)
-	require.NoError(t, err)
-	require.NotNil(t, ur)
-
-	en := NewAssignmentEntitlement(ur, "member", eopt.WithGrantableTo(rt))
-	require.NotNil(t, en)
-	require.Equal(t, v2.Entitlement_PURPOSE_VALUE_ASSIGNMENT, en.Purpose)
-	require.Equal(t, ur, en.Resource)
-	require.Equal(t, "member", en.DisplayName)
-	require.Equal(t, "member", en.Slug)
-	require.Len(t, en.GrantableTo, 1)
-	require.Equal(t, rt, en.GrantableTo[0])
-}
-
-func TestNewEntitlementID(t *testing.T) {
-	type args struct {
-		resource   *v2.Resource
-		permission string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			"ID for role member",
-			args{
-				resource: &v2.Resource{
-					Id: &v2.ResourceId{
-						ResourceType: "foo",
-						Resource:     "1234",
-					},
-				},
-				permission: "member",
-			},
-			"foo:1234:member",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := NewEntitlementID(tt.args.resource, tt.args.permission); got != tt.want {
-				t.Errorf("NewEntitlementID() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestNewGrant(t *testing.T) {
-	rt := NewResourceType("Group", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_GROUP})
-	ur, err := NewResource("test-group", rt, nil, 1234)
-	require.NoError(t, err)
-	require.NotNil(t, ur)
-
-	en := NewPermissionEntitlement(ur, "admin", eopt.WithGrantableTo(rt))
-	require.NotNil(t, en)
-
-	grant := NewGrant(ur, en.Slug, &v2.ResourceId{
-		ResourceType: "user",
-		Resource:     "567",
-	})
-	require.NotNil(t, grant)
-	require.NotNil(t, grant.Entitlement)
-	require.Equal(t, "group:1234:admin", grant.Entitlement.Id)
-	require.NotNil(t, grant.Principal)
-	require.Equal(t, "user", grant.Principal.Id.ResourceType)
-	require.Equal(t, "567", grant.Principal.Id.Resource)
-	require.Equal(t, "group:1234:admin:user:567", grant.Id)
 }
 
 func TestNewGroupResource(t *testing.T) {
@@ -121,7 +49,7 @@ func TestNewGroupResource(t *testing.T) {
 		"group_name": "Test",
 	}
 	rt := NewResourceType("Group", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_GROUP})
-	gr, err := NewGroupResource("test group", rt, nil, 1234, profile, ropt.WithAnnotation(&v2.V1Identifier{Id: "v1"}))
+	gr, err := NewGroupResource("test group", rt, nil, 1234, profile, WithAnnotation(&v2.V1Identifier{Id: "v1"}))
 	require.NoError(t, err)
 	require.NotNil(t, gr)
 	require.Equal(t, rt.Id, gr.Id.ResourceType)
@@ -135,35 +63,19 @@ func TestNewGroupResource(t *testing.T) {
 	require.True(t, ok)
 	require.Equal(t, "v1", v1ID.Id)
 
-	roleTrait, err := GetRoleTrait(gr)
+	roleTrait, err := sdk.GetRoleTrait(gr)
 	require.Error(t, err)
 	require.Nil(t, roleTrait)
-	groupTrait, err := GetGroupTrait(gr)
+	groupTrait, err := sdk.GetGroupTrait(gr)
 	require.NoError(t, err)
 	require.NotNil(t, groupTrait)
 	require.NotNil(t, groupTrait.Profile)
-	fName, foundProfileData := GetProfileStringValue(groupTrait.Profile, "group_name")
+	fName, foundProfileData := sdk.GetProfileStringValue(groupTrait.Profile, "group_name")
 	require.True(t, foundProfileData)
 	require.Equal(t, "Test", fName)
-	mName, foundProfileData := GetProfileStringValue(groupTrait.Profile, "first_name")
+	mName, foundProfileData := sdk.GetProfileStringValue(groupTrait.Profile, "first_name")
 	require.False(t, foundProfileData)
 	require.Equal(t, "", mName)
-}
-
-func TestNewPermissionEntitlement(t *testing.T) {
-	rt := NewResourceType("Group", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_GROUP})
-	ur, err := NewResource("test-group", rt, nil, 1234)
-	require.NoError(t, err)
-	require.NotNil(t, ur)
-
-	en := NewPermissionEntitlement(ur, "admin", eopt.WithGrantableTo(rt))
-	require.NotNil(t, en)
-	require.Equal(t, v2.Entitlement_PURPOSE_VALUE_PERMISSION, en.Purpose)
-	require.Equal(t, ur, en.Resource)
-	require.Equal(t, "admin", en.DisplayName)
-	require.Equal(t, "admin", en.Slug)
-	require.Len(t, en.GrantableTo, 1)
-	require.Equal(t, rt, en.GrantableTo[0])
 }
 
 func TestNewResource(t *testing.T) {
@@ -172,7 +84,7 @@ func TestNewResource(t *testing.T) {
 		Resource:     "567",
 	}
 	rt := NewResourceType("Role", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_ROLE})
-	rr, err := NewResource("test resource", rt, parentID, "1234", ropt.WithAnnotation(&v2.V1Identifier{Id: "v1"}))
+	rr, err := NewResource("test resource", rt, parentID, "1234", WithAnnotation(&v2.V1Identifier{Id: "v1"}))
 	require.NoError(t, err)
 	require.NotNil(t, rr)
 	require.Equal(t, rt.Id, rr.Id.ResourceType)
@@ -187,19 +99,19 @@ func TestNewResource(t *testing.T) {
 	require.True(t, ok)
 	require.Equal(t, "v1", v1ID.Id)
 
-	groupTrait, err := GetGroupTrait(rr)
+	groupTrait, err := sdk.GetGroupTrait(rr)
 	require.Error(t, err)
 	require.Nil(t, groupTrait)
 
-	userTrait, err := GetUserTrait(rr)
+	userTrait, err := sdk.GetUserTrait(rr)
 	require.Error(t, err)
 	require.Nil(t, userTrait)
 
-	roleTrait, err := GetRoleTrait(rr)
+	roleTrait, err := sdk.GetRoleTrait(rr)
 	require.Error(t, err)
 	require.Nil(t, roleTrait)
 
-	appTrait, err := GetAppTrait(rr)
+	appTrait, err := sdk.GetAppTrait(rr)
 	require.Error(t, err)
 	require.Nil(t, appTrait)
 }
@@ -217,7 +129,7 @@ func TestNewRoleResource(t *testing.T) {
 		"role_name": "Test",
 	}
 	rt := NewResourceType("Role", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_ROLE})
-	rr, err := NewRoleResource("test role", rt, nil, "1234", profile, ropt.WithAnnotation(&v2.V1Identifier{Id: "v1"}))
+	rr, err := NewRoleResource("test role", rt, nil, "1234", profile, WithAnnotation(&v2.V1Identifier{Id: "v1"}))
 	require.NoError(t, err)
 	require.NotNil(t, rr)
 	require.Equal(t, rt.Id, rr.Id.ResourceType)
@@ -231,18 +143,18 @@ func TestNewRoleResource(t *testing.T) {
 	require.True(t, ok)
 	require.Equal(t, "v1", v1ID.Id)
 
-	groupTrait, err := GetGroupTrait(rr)
+	groupTrait, err := sdk.GetGroupTrait(rr)
 	require.Error(t, err)
 	require.Nil(t, groupTrait)
 
-	roleTrait, err := GetRoleTrait(rr)
+	roleTrait, err := sdk.GetRoleTrait(rr)
 	require.NoError(t, err)
 	require.NotNil(t, roleTrait)
 	require.NotNil(t, roleTrait.Profile)
-	fName, foundProfileData := GetProfileStringValue(roleTrait.Profile, "role_name")
+	fName, foundProfileData := sdk.GetProfileStringValue(roleTrait.Profile, "role_name")
 	require.True(t, foundProfileData)
 	require.Equal(t, "Test", fName)
-	mName, foundProfileData := GetProfileStringValue(roleTrait.Profile, "first_name")
+	mName, foundProfileData := sdk.GetProfileStringValue(roleTrait.Profile, "first_name")
 	require.False(t, foundProfileData)
 	require.Equal(t, "", mName)
 }
@@ -254,7 +166,7 @@ func TestNewUserResource(t *testing.T) {
 		"last_name":  "User",
 	}
 	rt := NewResourceType("User", []v2.ResourceType_Trait{v2.ResourceType_TRAIT_USER})
-	ur, err := NewUserResource("test user", rt, nil, 1234, userEmail, profile, ropt.WithAnnotation(&v2.V1Identifier{Id: "v1"}))
+	ur, err := NewUserResource("test user", rt, nil, 1234, userEmail, profile, WithAnnotation(&v2.V1Identifier{Id: "v1"}))
 	require.NoError(t, err)
 	require.NotNil(t, ur)
 	require.Equal(t, rt.Id, ur.Id.ResourceType)
@@ -268,20 +180,20 @@ func TestNewUserResource(t *testing.T) {
 	require.True(t, ok)
 	require.Equal(t, "v1", v1ID.Id)
 
-	roleTrait, err := GetRoleTrait(ur)
+	roleTrait, err := sdk.GetRoleTrait(ur)
 	require.Error(t, err)
 	require.Nil(t, roleTrait)
 
-	ut, err := GetUserTrait(ur)
+	ut, err := sdk.GetUserTrait(ur)
 	require.NoError(t, err)
 	require.NotNil(t, ut)
 	require.Len(t, ut.Emails, 1)
 	require.Equal(t, userEmail, ut.Emails[0].Address)
 	require.NotNil(t, ut.Profile)
-	fName, foundProfileData := GetProfileStringValue(ut.Profile, "first_name")
+	fName, foundProfileData := sdk.GetProfileStringValue(ut.Profile, "first_name")
 	require.True(t, foundProfileData)
 	require.Equal(t, "Test", fName)
-	mName, foundProfileData := GetProfileStringValue(ut.Profile, "middle_name")
+	mName, foundProfileData := sdk.GetProfileStringValue(ut.Profile, "middle_name")
 	require.False(t, foundProfileData)
 	require.Equal(t, "", mName)
 }
